+++ apps/worxpace/src/app/(platform)/(tools)/workflow/[workflowId]/page.tsx
@@ -1,9 +1,0 @@
-const Workflow = () => {
-  return (
-    <div>
-      <h1>Workflow Editor</h1>
-    </div>
-  );
-};
-
-export default Workflow;